+++ src/compile/adapter/debug.js
@@ -1,9 +1,9 @@
 /**
- * 错误事件处理器
+ * 调试器
  * @param   {Object} error 
  * @returns {string}
  */
-const onerror = error => {
+const debug = error => {
     let output = `Template Error`;
     const message = JSON.stringify(error, null, 2);
 
@@ -14,4 +14,4 @@ const onerror = error => {
     return () => `{${output}}`;
 };
 
-module.exports = onerror;
+module.exports = debug;