+++ test/compile/adapter/onerror.js
@@ -1,15 +1,15 @@
 const assert = require('assert');
-const onerror = require('../../../src/compile/adapter/onerror');
+const debug = require('../../../src/compile/adapter/debug');
 
 
-describe('#compile/adapter/onerror', () => {
+describe('#compile/adapter/debug', () => {
 
     it('return function', () => {
-        assert.deepEqual('function', typeof onerror({}));
+        assert.deepEqual('function', typeof debug({}));
     });
 
     it('run function', () => {
-        assert.deepEqual('{Template Error}', onerror({})());
+        assert.deepEqual('{Template Error}', debug({})());
     });
 
 });